+++ geopackage-ios/core/contents/GPKGContents.m
@@ -50,7 +50,7 @@ NSString * const GPKG_CON_COLUMN_SRS_ID = @"srs_id";
 -(GPKGBoundingBox *) getBoundingBox{
     GPKGBoundingBox *boundingBox = nil;
     if(self.minX != nil && self.maxX != nil && self.minY != nil && self.maxY != nil){
-        boundingBox = [[GPKGBoundingBox alloc] initWithMinLongitude:self.minX andMaxLongitude:self.maxX andMinLatitude:self.minY andMaxLatitude:self.maxY];
+        boundingBox = [[GPKGBoundingBox alloc] initWithMinLongitude:self.minX andMinLatitude:self.minY andMaxLongitude:self.maxX andMaxLatitude:self.maxY];
     }
     return boundingBox;
 }